+++ packages/pi-coding-agent/src/utils/error.ts
@@ -1,0 +1,6 @@
+/**
+ * Extract a human-readable message from an unknown caught value.
+ */
+export function getErrorMessage(err: unknown): string {
+	return err instanceof Error ? err.message : String(err);
+}
